extract reducer/state to its own module
